add centralized header definitions
