+++ src/suite/request/mod.rs
@@ -1,6 +1,5 @@
 mod query_params;
 pub mod request_definition;
-pub mod request_headers;
 pub mod request_method;
 
 pub use request_definition::RequestDefinition;
