+++ src/suite/response/mod.rs
@@ -1,5 +1,4 @@
 pub mod response_definition;
-pub mod response_headers;
 pub mod status;
 
 pub use response_definition::ResponseDefinition;
