+++ src/suite/response/response_definition.rs
@@ -5,17 +5,17 @@ use serde::Deserialize;
 use crate::{
     errors::CaptiError,
     m_value::{m_match::MMatch, m_value::MValue, status_matcher::StatusMatcher},
-    suite::test::TestResult,
+    suite::{headers::MHeaders, test::TestResult},
     variables::{variable_map::VariableMap, SuiteVariables},
 };
 
-use super::{response_headers::ResponseHeaders, status::Status};
+use super::status::Status;
 
 #[derive(Debug, Clone, Deserialize, PartialEq)]
 pub struct ResponseDefinition {
     pub status: Status,
     #[serde(default)]
-    pub headers: ResponseHeaders,
+    pub headers: MHeaders,
     #[serde(default)]
     pub body: MValue,
 }
@@ -24,21 +24,13 @@ impl ResponseDefinition {
     pub async fn from_response(response: reqwest::Response) -> Self {
         let status = Status::from(StatusMatcher::Exact(response.status().as_u16()));
 
-        let headers = ResponseHeaders::from(response.headers());
+        let headers = MHeaders::from(response.headers());
 
         let body_text = response.text().await.unwrap_or("".to_string());
         let body = match serde_json::from_str::<MValue>(&body_text) {
             Ok(body) => body,
             Err(_) => MValue::String(body_text),
         };
-
-        // let body = match response.json::<MValue>().await {
-        //     Ok(body) => body,
-        //     Err(e) => {
-        //         let text_body = response.text().await.unwrap_or("".to_string());
-        //         MValue::String(text_body)
-        //     }
-        // };
 
         ResponseDefinition {
             status,
